refactor(web_core): move IndexImplementation to v1_0/functions/system_functions
